+++ src/guest.rs
@@ -1,0 +1,5 @@
+//! Guest pointer width which differs from host pointer width
+//! 
+//! Useful for developing hypervisors.
+
+// TODO: design this module